requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-hashid-field==3.1.3
hashids==1.3.1
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
django-heroku===0.3.1
psycopg2===2.8.6
gunicorn===20.0.4